myProject/requirements.txt
# === Django + Project Essentials ===
Django==5.1.2
dj-database-url==2.3.0
django-environ==0.10.0
gunicorn==20.1.0
whitenoise==6.7.0
sqlparse==0.5.1
#psycopg2==2.9.7         # PostgreSQL driver (perfect for Railway)
psycopg2-binary==2.9.7     
setuptools==75.2.0
asgiref==3.8.1
tzdata==2025.2          # updated from 2024.2 to match Celery requirement
packaging==24.1

# === Static + Media File Storage ===
pillow==10.4.0
cloudinary==1.43.0
django-cloudinary-storage==0.3.0

# === AI / Embedding Related ===
openai==1.51.0
faiss-cpu==1.10.0
numpy==2.2.4
pymupdf==1.25.5         # for PDF text extraction
tqdm==4.66.6            # for progress bars in scripts (optional)

# === Celery + Redis for Background Tasks ===
celery==5.5.0
redis==5.2.1
billiard==4.2.1
kombu==5.5.2
vine==5.1.0
amqp==5.3.1

# === HTTP & Cloud Handling ===
requests==2.32.3
certifi==2024.8.30
idna==3.10

# === Optional: Used by OpenAI or CLI dependencies ===
colorama==0.4.6
typing-extensions==4.12.2
httpx==0.27.2
httpcore==1.0.6
h11==0.14.0
anyio==4.5.2
sniffio==1.3.1
distro==1.9.0
jiter==0.7.0

# === Pydantic for data validation (used by OpenAI) ===
pydantic==2.9.2
pydantic-core==2.23.4
annotated-types==0.7.0

# === gdrive ===
cachetools==5.5.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.3
requests-oauthlib==2.0.0
rsa==4.9
uritemplate==4.1.1
